Refactor: 공통 페이지 구조 변경
