+++ src/app/(withHeader)/news/[id]/page.tsx
@@ -1,10 +1,10 @@
 import { cookies } from 'next/headers';
 
 import { fetcher } from '@/api/fetcher';
-import NotFound from '@/components/common/NotFound';
-import RedirectWithModal from '@/components/common/RedirectWithModal';
 import { ImageViewer } from '@/components/molecules/ImageViewer';
 import { MarkdownViewer } from '@/components/molecules/MarkdownViewer';
+import { NotFoundPage } from '@/components/pages/NotFoundPage';
+import { RedirectWithLoginModalPage } from '@/components/pages/RedirectWithLoginModalPage';
 import PostCommentForm from '@/components/post/PostCommentForm';
 import PostCommentList from '@/components/post/PostCommentList';
 import PostHeader from '@/components/post/PostHeader';
@@ -31,7 +31,7 @@ export default async function NewsDetailPage({ params }: NewsDetailPageProps) {
     });
 
     if (res?.code === 401 || res?.code === 403) {
-      return <RedirectWithModal />;
+      return <RedirectWithLoginModalPage />;
     }
 
     if (res) news = res.data;
@@ -39,7 +39,7 @@ export default async function NewsDetailPage({ params }: NewsDetailPageProps) {
     console.error('fetch failed:', e);
   }
 
-  if (!news) return <NotFound />;
+  if (!news) return <NotFoundPage />;
 
   return (
     <div className="flex justify-center items-center w-full">
