+++ src/app/(withHeader)/notices/[id]/page.tsx
@@ -1,10 +1,10 @@
 import { cookies } from 'next/headers';
 
 import { fetcher } from '@/api/fetcher';
-import NotFound from '@/components/common/NotFound';
-import RedirectWithModal from '@/components/common/RedirectWithModal';
 import { ImageCarousel } from '@/components/molecules/ImageCarousel';
 import { MarkdownViewer } from '@/components/molecules/MarkdownViewer';
+import { NotFoundPage } from '@/components/pages/NotFoundPage';
+import { RedirectWithLoginModalPage } from '@/components/pages/RedirectWithLoginModalPage';
 import PostCommentForm from '@/components/post/PostCommentForm';
 import PostCommentList from '@/components/post/PostCommentList';
 import PostFileItem from '@/components/post/PostFileItem';
@@ -32,7 +32,7 @@ export default async function NoticeDetailPage({ params }: NoticeDetailPageProps
     });
 
     if (res?.code === 401 || res?.code === 403) {
-      return <RedirectWithModal />;
+      return <RedirectWithLoginModalPage />;
     }
 
     if (res) notice = res.data;
@@ -40,7 +40,7 @@ export default async function NoticeDetailPage({ params }: NoticeDetailPageProps
     console.error('fetch failed:', e);
   }
 
-  if (!notice) return <NotFound />;
+  if (!notice) return <NotFoundPage />;
 
   return (
     <div className="flex justify-center items-center w-full">
