+++ src/components/pages/NotFoundPage/index.ts
@@ -1,0 +1,1 @@
+export * from './NotFoundPage';
