+++ src/components/pages/RedirectWithLoginModalPage/index.ts
@@ -1,0 +1,1 @@
+export * from './RedirectWithLoginModalPage';
